+++ packages/server/utils/OpenAIServerManager.ts
@@ -1,5 +1,4 @@
 import OpenAI from 'openai'
-import type {ChatCompletionCreateParamsNonStreaming} from 'openai/resources'
 import {ModifyType} from '../graphql/public/resolverTypes'
 import {RetroReflection} from '../postgres/types'
 import {Logger} from './Logger'
@@ -11,7 +10,7 @@ type InsightResponse = {
 }
 
 class OpenAIServerManager {
-  private openAIApi
+  openAIApi
   constructor() {
     if (!process.env.OPEN_AI_API_KEY) {
       this.openAIApi = null
@@ -23,16 +22,6 @@ class OpenAIServerManager {
     })
   }
 
-  async chatCompletion(body: ChatCompletionCreateParamsNonStreaming) {
-    if (!this.openAIApi) return null
-    try {
-      return await this.openAIApi.chat.completions.create(body)
-    } catch (e) {
-      const error = e instanceof Error ? e : new Error('OpenAI failed to getSummary')
-      sendToSentry(error)
-      return null
-    }
-  }
   async getStandupSummary(
     responses: Array<{content: string; user: string}>,
     meetingPrompt: string